Create watch.py
a variation of game.py from targetgame.

for touch sensor on gpio pin.
watch for input change, then watch to see if the input is a double tap. 

single tap: cancel
double tap: turn on relay until the failsafe timer kills it or a single tap.
press and hold: turn on relay until released.